laptop/requirements.txt
ultralytics
flask
flask-cors
supabase
requests
Pillow
python-dotenv
streamlit
streamlit-autorefresh
numpy
opencv-python
